task #43 #44: S'ha millorat el html i l'aspecte en general de les vistes, aixi com s'han arreglat errors generals del programa

+++ HackerNews/vote/views.py
@@ -18,30 +18,38 @@ def vote_submission(request, id):
             print("incorrect")
         return redirect(request.META.get('HTTP_REFERER'))
     else:
-        return redirect("/login")
+        return redirect("Login")
 
 
 def unvote_submission(request, id):
-    subm = Submission.objects.get(id=id)
-    subm.points -= 1
-    subm.save()
-    act = ActionType.objects.get(name="Submission")
-    Vote.objects.get(submission=subm,user=request.user,type=act).delete()
-    return redirect(request.META.get('HTTP_REFERER'))
+    if request.user.is_authenticated:
+        subm = Submission.objects.get(id=id)
+        subm.points -= 1
+        subm.save()
+        act = ActionType.objects.get(name="Submission")
+        Vote.objects.get(submission=subm,user=request.user,type=act).delete()
+        return redirect(request.META.get('HTTP_REFERER'))
+    else:
+        return redirect("Login")
 
 
 def vote_comment(request,idc):
-
-    comm = Comment.objects.get(id=idc)
-    act = ActionType.objects.get(name="Comment")
-    vote = Vote.objects.get_or_create(comment=comm, user=request.user, type=act)
-    return redirect(request.META.get('HTTP_REFERER'))
+    if request.user.is_authenticated:
+        comm = Comment.objects.get(id=idc)
+        act = ActionType.objects.get(name="Comment")
+        vote = Vote.objects.get_or_create(comment=comm, user=request.user, type=act)
+        return redirect(request.META.get('HTTP_REFERER'))
+    else:
+        return redirect("Login")
 
 
 
 
 def unvote_comment(request,idc):
-    comm = Comment.objects.get(id=idc)
-    act = ActionType.objects.get(name="Comment")
-    Vote.objects.get(comment=comm, user=request.user,type=act).delete()
-    return redirect(request.META.get('HTTP_REFERER'))
+    if request.user.is_authenticated:
+        comm = Comment.objects.get(id=idc)
+        act = ActionType.objects.get(name="Comment")
+        Vote.objects.get(comment=comm, user=request.user,type=act).delete()
+        return redirect(request.META.get('HTTP_REFERER'))
+    else:
+        return redirect("Login")
